Split extwatflux_ss class:
Splitted the extwatflux_ss class based on coupler calls.
Methods called by decl are now in a separate file.
Methods called by runtimestart are in a separate file.
